playout-light.pl: use files from MediaFiles
In the past, shows and their audio files were mapped to each other.
This allowed to schedule repeating shows without need to re-upload
files.  With this change just all local audio files are played.  To
repeat a show one either uploads the same audio file to another event or
just create a local link in the filesystem.